refactor: store monetary values as integer cents and update README for v2.2

+++ expense.h
@@ -15,7 +15,7 @@ struct Expense
     int id;
     char description[50];
     char category[50];
-    float amount;
+    long amount;
     struct Date date;
 };
